requirements.txt
# 核心測試框架
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-xdist==3.5.0
pytest-cov==4.1.0
pytest-html==4.1.1

# UI 測試
playwright==1.40.0

# API 測試
requests==2.31.0
httpx==0.25.2

# API 測試工具 (符合 JD 要求)
pytest-requests==0.1.1
requests-toolbelt==1.0.0

# 效能測試 (JMeter 整合)
locust==2.20.0  # Python 版的效能測試工具，可替代 JMeter
pytest-benchmark==4.0.0  # 效能基準測試

# AI/ML 測試
transformers==4.35.2
torch==2.9.0
nltk==3.8.1
spacy==3.7.2

# 資料處理
pandas==2.1.4
numpy==1.26.2

# 測試報告
allure-pytest==2.13.2

# 工具
python-dotenv==1.0.0
pyyaml==6.0.1

# 監控與可觀測性
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-prometheus==0.59b0
opentelemetry-instrumentation-requests==0.59b0
prometheus-client==0.19.0